Revert "조인환[fix]: 전자문서 리스트호출 , 상세보기, 승인&반려 처리 쿼리 수정 외 "
