--- src/main/java/com/ga/gaent/mapper/ChatMapper.java
@@ -1,12 +1,0 @@
-package com.ga.gaent.mapper;
-
-import java.util.List;
-import org.apache.ibatis.annotations.Mapper;
-import com.ga.gaent.dto.ChatDTO;
-
-@Mapper
-public interface ChatMapper {
-    
-    void save(ChatDTO message);
-    List<ChatDTO> findByRoomId(String roomId);
-}
