--- src/main/java/com/ga/gaent/mapper/EdocProcessMapper.xml
@@ -1,218 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.ga.gaent.mapper.EdocProcessMapper">
-
-
-
-    <!--
-        @author : 정건희,조인환
-        Description : 전자결재문서 작성 후 등록할 때 입력되는 정보
-    -->
-    <insert id="insertEdoc" parameterType="com.ga.gaent.dto.EdocRequestDTO" useGeneratedKeys="true" keyProperty="edocNum">
-        INSERT INTO
-            edoc(
-                edoc_title,
-                edoc_type,
-                edoc_writer,
-                edoc_enroll_date
-            ) VALUES(
-                #{edocTitle},
-                #{edocType},
-                #{edocWriter},
-                NOW()
-            )
-    </insert>
-
-    <!--
-        @author : 정건희
-        Description : 전자결재문서 작성 후 등록할 때 입력되는 첨부파일의 정보
-    -->
-    <insert id="insertEdocFile" parameterType="Map">
-        INSERT INTO
-            edoc_file(
-                file_name,
-                original_file_name,
-                file_type,
-                file_size,
-            ) VALUES(
-                #{fileName},
-                #{orginalFileName},
-                #{fileType},
-                #{fileSize}
-            )
-    </insert>
-    
-
-
-    <!--
-        @author : 정건희
-        Description : 전자결재문서 작성 후 등록할 때 입력되는 결재선(1차, 2차)의 정보
-    -->
-    <insert id="insertApprover" parameterType="Map">
-        INSERT INTO
-            edoc_approval(
-                edoc_num,
-                approver,
-                appr_order
-            ) VALUES(
-                #{edocNum},
-                #{approver},
-                #{apprOrder}
-            )
-    </insert>
-    
-    
-    
-
-    <!--
-        @author : 정건희
-        Description : 전자결재문서 작성 후 등록할 때 입력되는 기안서 양식의 정보
-    -->
-    <insert id="insertEdocDraft" parameterType="com.ga.gaent.dto.EdocFormTypeDTO">
-        INSERT INTO
-            edoc_draft(
-                edoc_num,
-                prop_start_date,
-                prop_end_date,
-                prop_content
-            ) VALUES(
-                #{edocNum},
-                #{propStartDate},
-                #{propEndDate},
-                #{propContent}
-            )
-    </insert>
-
-     <!-- 휴가신청서  -->
-    <insert id="insertEdocVacation" parameterType="com.ga.gaent.dto.EdocFormTypeDTO">
-        INSERT INTO
-            edoc_vacation(
-                edoc_num,
-                vac_type,
-                vac_start_date,
-                vac_end_date,
-                vac_reason
-            ) VALUES(
-                #{edocNum},
-                #{vacType},
-                #{vacStartDate},
-                #{vacEndDate},
-                #{vacReason}
-            )
-    </insert>
-    
-    <!-- 지출결의서  -->
-    <insert id="insertEdocProject" parameterType="com.ga.gaent.dto.EdocFormTypeDTO">
-        INSERT INTO
-            edoc_project(
-                edoc_num,
-                project_content,
-                project_estimate
-            ) VALUES(
-                #{edocNum},
-                #{projectContent},
-                #{projectEstimate}
-            )
-    </insert>
-    
-    <!-- 경조사  -->
-    <insert id="insertEdocEvent" parameterType="com.ga.gaent.dto.EdocFormTypeDTO">
-        INSERT INTO
-            edoc_event(
-                edoc_num,
-                event_type,
-                event_date,
-                event_place,
-                event_expense
-            ) VALUES(
-                #{edocNum},
-                #{eventType},
-                #{eventDate},
-                #{eventPlace},
-                #{eventExpense}
-            )
-    </insert>
-
-    <!-- 보고서  -->
-    <insert id="insertEdocReport" parameterType="com.ga.gaent.dto.EdocFormTypeDTO">
-        INSERT INTO
-            edoc_report(
-                edoc_num,
-                report_type,
-                report_content
-                ) 
-            VALUES(
-                #{edocNum},
-                #{reportType},
-                #{reportContent}
-                )
-    </insert>
-
-    
-    <!-- 결재, 반려처리 -->
-    <!-- 결재선 -->
-    <update id="updateEdocApprovalStatus" parameterType="map">
-        UPDATE 
-            edoc_approval
-        SET 
-            appr_reason = #{apprReason},
-            appr_status = #{apprStatus},
-            appr_date = NOW()
-        WHERE 
-            edoc_num = #{edocNum} AND
-            approver = #{empCode}
-    </update>
-    
-    
-   
-           
-    <!-- 첫번째 업데이트 -->
-    <update id="updateEdocStatusFirst" parameterType="map">
-        UPDATE  
-            edoc
-        <set>
-            edoc_status = #{request},
-            <if test="request == -1">
-                edoc_done_date = NOW(),
-            </if>
-            edoc_enroll_date = NOW()
-        </set>
-        WHERE 
-            edoc_num = #{edocNum} 
-    </update>
-    
-    
-    <!--  두번째 업데이트  -->
-    <update id="updateEdocStatusSecond" parameterType="map">
-        UPDATE  
-            edoc
-        <set>
-            edoc_status = #{request},
-            edoc_done_date = NOW(),
-            edoc_enroll_date = NOW()
-        </set>
-        WHERE 
-            edoc_num = #{edocNum} 
-    </update>
-    
-    
-   
-    
-    
-    <select id="checkEdocStatus" parameterType="String">
-        SELECT 
-            edoc_status
-        FROM 
-            edoc
-        WHERE 
-            edoc_num = #{edocNum}
-    </select>
-    
-    
-    
-    
-    
-    
-    
-
-</mapper>
