+++ src/main/resources/static/assets/css/sidebar.css
@@ -6,5 +6,5 @@
 
 .subside-nav {
   margin-top: 58px;
-  box-shadow: 0 0 0.125rem 0 rgba(67, 89, 113, 0.1);
+  box-shadow: 0 0.125rem 1rem 0 rgba(67, 89, 113, 0.1);
 }
