+++ src/main/resources/static/scss/_bootstrap-extended/_tables.scss
@@ -11,7 +11,6 @@
 .card {
   .table {
     margin-bottom: 0;
-    left: calc(-21rem - 1rem);
   }
 }
 @supports (-moz-appearance: none) {
