--- src/main/webapp/WEB-INF/view/chat/chat-sub-sidebar.jsp
@@ -1,27 +1,0 @@
-<%@ page language="java" contentType="text/html; charset=UTF-8" pageEncoding="UTF-8" %>
-<body>
-  <aside id="layout-menu" class="layout-chat-menu menu-vertical menu bg-sub-menu-theme subside-nav">
-    <div class="menu-inner-shadow"></div>
-    <ul class="menu-inner py-3">
-      <li class="menu-item">
-        <div class="menu-link">
-          <span class="display-6 fw-semibold mb-0">직원목록</span>
-        </div>
-      </li>
-      <li class="menu-item">
-        <div class="menu-link">
-        </div>
-      </li>
-      <li class="menu-item">
-        <div class="menu-link">
-          <div class="d-flex align-items-center">
-            <input class="form-control me-2" type="text" name="searchEmp">
-            <button type="button" class="btn btn-icon btn-outline-primary">
-              <span class="tf-icons bx bx-search"></span>
-            </button>
-          </div>
-        </div>
-      </li>
-    </ul>
-  </aside>
-</body>
